requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
bracex==2.6
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
deepagents==0.4.1
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.3
filetype==1.2.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-core==1.2.11
langchain-google-genai==4.2.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.4.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typer-slim==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
watchdog==6.0.0
wcmatch==10.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
